requirements.txt
Flask
RPi.GPIO
sqlite3